refactor: extract more into `helpers.py`
